Création du component InputSteps pour simplication du composant FormCreateRecipe
