+++ frontend/src/pages/user/pages/CreateRecipe/components/formCreateRecipe/FormCreateRecipe.jsx
@@ -4,7 +4,7 @@ import { yupResolver } from '@hookform/resolvers/yup';
 import { getAllIngredients, createRecipeApi } from '../../../../../../api';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
-  faUtensils, faImage,
+  faUtensils,
   faListUl, faPlus, faMinus, faListOl, faTrash
 } from '@fortawesome/free-solid-svg-icons';
 import { UserContext } from '../../../../../../contexts/UserContext';
@@ -13,6 +13,7 @@ import SelectCookingTime from './components/select/selectTCookIngTime/SelectCook
 import SelectLevel from './components/select/selectLevel/SelectLevel';
 import InputTitle from './components/inputTitle/InputTitle';
 import InputImage from './components/inputImage/InputImage';
+import InputSteps from './components/inputSteps/InputSteps';
 
 
 // Définir le schéma de validation avec yup
@@ -21,9 +22,8 @@ const FormCreateRecipe = () => {
   const { user } = useContext(UserContext);
   console.log("user in FormCreateRecipe :", user);
   const [ingredients, setIngredients] = useState([]);
-  const [selectedIngredients, setSelectedIngredients] = useState([]);
   const [steps, setSteps] = useState([]);
-  const [currentStep, setCurrentStep] = useState('');
+  const [selectedIngredients, setSelectedIngredients] = useState([]);
   const [error, setError] = useState(null);
 
   // Initialiser les fonctions du formulaire avec react-hook-form
@@ -60,17 +60,6 @@ const FormCreateRecipe = () => {
   };
 
   // Ajouter une étape à la liste des étapes
-  const handleAddStep = () => {
-    if (currentStep.trim() !== '') {
-      setSteps([...steps, currentStep.trim()]);
-      setCurrentStep('');
-    }
-  };
-
-  // Supprimer une étape de la liste des étapes
-  const handleRemoveStep = (index) => {
-    setSteps(steps.filter((_, i) => i !== index));
-  };
 
   // Filtrer les options d'ingrédients en fonction de ceux qui sont déjà sélectionnés
   const filteredIngredientOptions = ingredients
@@ -83,18 +72,6 @@ const FormCreateRecipe = () => {
 
 
   // Convertir les étapes en éléments de liste
-  const stepsList = steps.map((step, index) => (
-    <li key={index} className="flex items-center justify-between bg-orange-50 p-2 rounded">
-      <span>{step}</span>
-      <button
-        type="button"
-        onClick={() => handleRemoveStep(index)}
-        className="text-red-500 hover:text-red-700"
-      >
-        <FontAwesomeIcon icon={faTrash} />
-      </button>
-    </li>
-  ));
 
   // Convertir les ingrédients sélectionnés en éléments d'affichage
   const selectedIngredientsList = selectedIngredients.map(ingredient => (
@@ -118,7 +95,7 @@ const FormCreateRecipe = () => {
       setError("Utilisateur non connecté.");
       return;
     }
-  
+
     if (selectedIngredients.length === 0) {
       setError("Veuillez sélectionner au moins un ingrédient.");
       return;
@@ -127,14 +104,14 @@ const FormCreateRecipe = () => {
       setError("Veuillez ajouter au moins une étape.");
       return;
     }
-  
+
     const recipeData = {
       ...data,
       ingredients: selectedIngredients.map(ing => ({ _id: ing._id })),  // Vérifiez que les IDs sont correctement extraits
       steps: steps,
       creator: user._id,
     };
-  
+
     try {
       await createRecipeApi(recipeData);
       console.log('Recette créée avec succès !');
@@ -143,7 +120,7 @@ const FormCreateRecipe = () => {
       setError("Impossible de créer la recette. Veuillez réessayer plus tard.");
     }
   };
-  
+
   return (
     <div className="max-w-4xl mx-auto bg-orange-100 p-8 rounded-lg shadow-xl">
       <h2 className="text-3xl font-bold mb-6 text-orange-800 text-center">Créer une nouvelle recette</h2>
@@ -165,12 +142,12 @@ const FormCreateRecipe = () => {
         </div>
 
         <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-          <SelectCookingTime 
-            register={register} 
-            errors={errors} 
+          <SelectCookingTime
+            register={register}
+            errors={errors}
           />
-          
-         <SelectLevel register={register} errors={errors} />
+
+          <SelectLevel register={register} errors={errors} />
         </div>
 
         <div>
@@ -190,35 +167,7 @@ const FormCreateRecipe = () => {
         <div className="grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-4">
           {selectedIngredientsList}
         </div>
-
-        <div>
-          <label htmlFor="steps" className="flex items-center text-sm font-medium text-orange-700">
-            <FontAwesomeIcon icon={faListOl} className="mr-2" />
-            Étapes de préparation
-          </label>
-          <div className="flex mt-1">
-            <input
-              type="text"
-              value={currentStep}
-              onChange={(e) => setCurrentStep(e.target.value)}
-              className="flex-grow rounded-l-md border-orange-300 focus:border-orange-500 focus:ring focus:ring-orange-200 focus:ring-opacity-50"
-              placeholder="Ajouter une étape"
-            />
-            <button
-              type="button"
-              onClick={handleAddStep}
-              className="bg-orange-500 text-white px-4 py-2 rounded-r-md hover:bg-orange-600 transition duration-300"
-            >
-              <FontAwesomeIcon icon={faPlus} />
-            </button>
-          </div>
-          {steps.length > 0 && (
-            <ol className="list-decimal list-inside mt-2 space-y-2">
-              {stepsList}
-            </ol>
-          )}
-        </div>
-
+          <InputSteps register={register} errors={errors} steps={steps} setSteps={setSteps} />
         <div>
           <button type="submit" className="w-full bg-orange-600 text-white px-4 py-2 rounded-md hover:bg-orange-700 focus:outline-none focus:ring-2 focus:ring-orange-500 focus:ring-opacity-50 transition duration-300 flex items-center justify-center">
             <FontAwesomeIcon icon={faUtensils} className="mr-2" />
